regulator: tps62864: add roadtest
Add a roadtest for the recently-added tps62864 regulator driver.  It
tests voltage setting, mode setting, as well as devicetree mode
translation.  It uses the recently-added devicetree support in
regulator-virtual-consumer.

All the variants supported by the driver have identical register
interfaces so only one test/model is added.

It requires the following patches which are, as of writing, not in
mainline:

 - regulator: Add support for TPS6286x
   https://lore.kernel.org/lkml/20220204155241.576342-3-vincent.whitchurch@axis.com/

 - regulator: virtual: add devicetree support
   https://lore.kernel.org/lkml/20220301111831.3742383-4-vincent.whitchurch@axis.com/

Signed-off-by: Vincent Whitchurch <vincent.whitchurch@axis.com>