roadtest: add build files
Add a Makefile and other miscellaneous build-related files for the
roadtest framework.

To make it easier to run the tests on systems which do not have the
required libraries or Python version, a Dockerfile is included and the
Makefile has built-in support for running the tests in a Docker
container.

Targets for code formatting and static checking of the Python code are
included.

Signed-off-by: Vincent Whitchurch <vincent.whitchurch@axis.com>