regulator: virtual: warn against production use
This driver is only meant for debugging and testing.  Currently, it's
not possible to use it without patching the kernel since it requires
platform data, but we'll be adding devicetree support, so add a loud
warning to make it clear that it's still only meant for debugging and
testing.

Signed-off-by: Vincent Whitchurch <vincent.whitchurch@axis.com>
Link: https://lore.kernel.org/r/20220301111831.3742383-3-vincent.whitchurch@axis.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit d2fb5487ecb2a28b61ff261ae18488afc98d24a6)
Signed-off-by: Vincent Whitchurch <vincent.whitchurch@axis.com>